Update add_group_modal to match new file names.
